feat: Major cleanup of code and changes to file structure
The purpose of this commit is to delinate the functions into their proper
files (such as all logging functions have been moved to the logging files)
and to clean up existing code in preparation for the integration of more than
two edge cameras. Certain bug fixes were also made that previously allowed for
the pairing of new tracks to existing paired ones. Redundant functions and operations
removed. Added the feature to turn on/off certain information display in the UI.
Most constants parameterised. Creation of a new matched_tracks map to keep track
of which cameras the matched tracks appear in.