fix: Temporary fix for SRT and websocket sync issues
This is a temporary fix for the timesync issues that cause the incoming detections
from the vilota websocket to become intermittent, thus affecting the tracking ability.
The detections are forced to "stick" around longer, by not clearing them when there are
zero detections received from the vilota websocket.

The fix will be removed when the timesync issue is resolved.